+++ programmers/REACT-TASK-APP/src/components/BoardList/SideForm/SideForm.tsx
@@ -1,5 +1,10 @@
 import { ChangeEvent, useState } from "react"
 import { FiCheck } from "react-icons/fi"
+import { icon, input, sideForm } from "./SideForm.css"
+import { useTypedDispatch } from "../../../hooks/redux"
+import { addBoard } from "../../../store/slices/boardsSlice"
+import { v4 as uuidv4 } from 'uuid'
+import { addLog } from "../../../store/slices/loggerSlice"
 
 interface Props {
   setIsFormOpen: React.Dispatch<React.SetStateAction<boolean>>  
@@ -8,6 +13,7 @@ interface Props {
 
 const SideForm = ({ setIsFormOpen, inputRef }: Props) => {
   const [inputText, setInputText] = useState('1')
+  const dispatch = useTypedDispatch();
 
   const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
     setInputText(e.target.value)
@@ -17,16 +23,34 @@ const SideForm = ({ setIsFormOpen, inputRef }: Props) => {
     setIsFormOpen(false)
   }
 
+  const handleClick = () => {
+    if (inputText) {
+      dispatch(
+        addBoard({board: {boardId: uuidv4(), boardName: inputText, lists:[] }})
+      )
+
+      dispatch(
+        addLog({
+          logId: uuidv4(),
+          logMessage: `게시판 등록: ${inputText}`,
+          logAuthor: 'User',
+          logTimeStamp: String(Date.now())
+        })
+      )
+    }
+  }
+
   return (
-    <div>
+    <div className={sideForm}>
       <input type="text" 
         ref={inputRef}
+        className={input}
         placeholder="새로운 게시판 등록하기"
         value={inputText}
         onChange={handleChange}
         onBlur={handleOnBlur}
       />
-      <FiCheck />
+      <FiCheck className={icon} onMouseDown={handleClick}/>
     </div>
   )
 }